pyproject.toml
[project]
name = "memoryguard-ai"
version = "0.1.0"
description = "LLM Agent long-term memory poisoning detector"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0"
]

[project.scripts]
memoryguard = "memoryguard.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
